encryption and decryption
A software house is looking for a programmer to write a program that does encryption and decryption
of a message to securely transmit it over the network. Be the one who will write that program. Write a
C program that takes a string as an input (including spaces). The string will be passed to a function
named encrypt () as an argument. The function will encrypt the string and displays the encrypted
version of the message. There should be another function named decrypt (), which will decrypt the
message and displays the original message. Your program should be menu driven.
Implement Caesar cipher algorithm in a way that each character in the
original string should be replaced by its succeeding third letter.