Write a C program that does the following according to user input.
It’s a
menu-driven program. The user is asked to enter a string. Then the program
displays the following menu:
A) Count the number of vowels in the string
B) Count both the vowels and consonants in the string
C) Display the most frequent character in the string.
D) Concatenate another string with the existing string.
E) Exit the program.
The program performs the operation selected by the user and repeats until
the user selects E to exit the program.