Write a program that simulates the rolling of two dice.
The program should
use rand to roll the first die and should use rand again to roll the second die. The sum of the
two values should then be calculated. [Note: Each die can show an integer value from 1 to 6, so
the sum of the two values will vary from 2 to 12, with 7 being the most frequent sum and 2 and
12 being the least frequent sums.] Figure shows the 36 possible combinations of the two dice.
Your program should roll the two dice 1,000 times. Use a one-dimensional array to tally the
numbers of times each possible sum appears. Print the result count.