Now write a program that obtains a cricket match score and remaining wickets from user.
Suppose there is a cricket match between Pakistan and India. India played first and gave a
target of 300 runs to Pakistan If the score is greater than 300 outputs a message indicating
“Pakistan win the match by X wickets” while X are remaining wickets. Otherwise shows a
message how many remaining runs required to win with X wickets in hand. Use single if-else
statement. If score is 300 with 0 wickets in hand then display match is draw.
Note: Score and wickets are not less then zero