A program uses a char variable named membership and an int variable named age.
The
membership variable contains one of the following letters (entered in either uppercase or
lowercase): M or N. The letter M stands for member, and the letter N stands for non- member.
The program should display the appropriate seminar fee, which is based on a person’s
membership status and age. The fee schedule is shown in Figure. Write the C code to display
the fee.
$10 Club member less than 65 years old
$5 Club member atleast 65 years old
$20 Non Member