In this problem, the program starts up assuming that it is following an object on a 2- dimensional plane as it takes many steps to perform a sort‐of “random walk”
The object starts off at the “origin” (x = 0
and y = 0). You will input two real numbers that will represent the next coordinates of the object. The
computer will calculate the straight‐line distance that it has to travel to get there and position the
object at that point. Then you will input two more real numbers representing a move from that
location, then another position (another pair of points) and another position (pair of points) and so on.
To stop the program, you have to input the identical new coordinates as the previous step—that is,
there is no distance to be traveled for that final step. This is a “sentinel” indicating that the loop should
terminate. This last zero-step should not be counted in the number of steps. When the program exits
the loop at this point make the program print out the following points.
1. Report the total distance travelled.
2. Report the total number of steps.
3. Report the average distance traveled per step. (Make sure that the case of zero steps and zero
distance traveled does not cause problems.)
Use the following formula to calculate the distance of the step.
s=sqrt((xi-x0)^2+(yi-y0)^2)
Here,
xi, yi represent (New coordinates),
x0, y0 represent (Current coordinates) and
s is the Distance of a single step